% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLT_exponential_movie.R
\name{clt_exp}
\alias{clt_exp}
\title{Central Limit Theorem movie: exponential data}
\usage{
clt_exp(n = 30, lambda = 1, xlab = "x", pos = 1,
  envir = as.environment(pos))
}
\arguments{
\item{n}{An integer scalar.  The size of the samples drawn from a
normal distribution.}

\item{lambda}{A numeric scalar.  The rate parameter of the exponential
distribution from which data are to be simulated using \code{\link{rexp}}.}

\item{xlab}{A character scalar.  A name to use to label the horizontal
axis of the plots.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate the idea of a sampling distribution and the central
limit theorem (CLT) in a situation where data are simulated randomly
from an exponential distribution.
}
\details{
Loosely speaking, a consequence of the
  \href{https://en.wikipedia.org/wiki/Central_limit_theorem}{Central Limit Theorem (CLT)}
  is that, in many situations, the mean of a \strong{large number} of
  independent random variables has \strong{approximately} a normal distribution,
  even if these original variables are not normally distributed.

  This movie illustrates this in the case where the original variables
  are exponentially distributed.  Samples of size \code{n} are repeatedly
  simulated from an exponential distribution.  These samples are
  summarized using a histogram that appears at the top of the movie screen.
  For each sample the mean of these \code{n} values is calculated, stored
  and added to another histogram plotted below the first histogram.
  The (exponential) probability density function (p.d.f.) of the original
  variables is superimposed on the top histogram.  On the bottom histogram
  is superimposed the approximate (large \code{n}) p.d.f. given by the
  CLT.

  The user may choose the sample size \code{n}, that is, the number of
  values over which a mean is calculated, the rate parameter \code{lambda}
  of the exponential normal distribution from which values are simulated
  and the label \code{xlab} for the horizontal axis.

  Once it starts, two aspects of this movie are controlled by the user.
  Firstly, there are buttons to increase (+) or decrease (-) the sample
  size, that is, the number of values over which a mean is calculated.
  Then there is a button labelled "simulate another sample of size n".
  Each time this button is clicked a new sample is simulated and its sample
  mean added to the bottom histogram.

  Another movie (\code{\link{clt_norm}}) considers the special case
  where the original variables are normally distributed.
}
\examples{
\dontrun{
clt_exp()
clt_exp(n = 10)
}
}
\seealso{
\code{\link{smovie}}: general information about smovie.

\code{\link{clt_norm}}: a similar movie using data simulated from a
  normal distribution.
}
