% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{discrete}
\alias{discrete}
\title{Univariate Discrete Distributions: p.m.f and c.d.f.}
\usage{
discrete(distn, params = list(), plot_par = list(), panel_plot = TRUE,
  hscale = NA, vscale = hscale, observed_value = NA, ...)
}
\arguments{
\item{distn}{A character scalar specifying the distribution from which
  observations are sampled..   Distributions \code{"binomial"},
  \code{"geometric"}, \code{"hypergeometric"}, \code{"negative binomial"},
  \code{Poisson} are recognised, case being ignored.
  The relevant distributional functions \code{dxxx} and \code{pxxx} in the
  \code{\link[stats]{stats-package}} are used.  The abbreviations
  \code{xxx} are also recognised.

  If \code{distn} is not supplied then \code{distn = "binomial"}
  is used.}

\item{params}{A named list of additional arguments to be passed to the
  density function associated with distribution \code{distn}.

  If a parameter value is not supplied then the default values in the
  relevant distributional function set using \code{distn} are used}

\item{plot_par}{A named list of graphical parameters
(see \code{link[graphics]{par}}) to be passed to
\code{\link[graphics]{plot}}.}

\item{panel_plot}{A logical parameter that determines whether the plot
is placed inside the panel (\code{TRUE}) or in the standard graphics
window (\code{FALSE}).  If the plot is to be placed inside the panel
then the tkrplot library is required.}

\item{hscale, vscale}{Numeric scalars.  Scaling parameters for the size
of the plot when \code{panel_plot = TRUE}. The default values are 1.4 on
Unix platforms and 2 on Windows platforms.}

\item{observed_value}{A non-negative integer.  If \code{observed_value} is
supplied then the corresponding line in the plot of the p.m.f. is coloured
in red.}

\item{...}{Additional arguments to be passed to
\code{\link[rpanel]{rp.doublebutton}}, not including \code{panel},
\code{variable}, \code{title}, \code{step}, \code{action}, \code{initval},
\code{range}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate how the probability mass function (p.m.f.) and
cumulative distribution function (c.d.f.) of a discrete random variable
depends on the values of its parameters.
}
\details{
Ass details.
}
\examples{
\dontrun{
discrete()
}
}
\seealso{
\code{\link{smovie}}: general information about smovie.
}
