% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_pearson.R
\name{FPearson}
\alias{FPearson}
\alias{dFPearson}
\title{Fisher's transformation of the Pearson produce moment correlation
coefficient}
\usage{
dFPearson(x, N, rho = 0, log = FALSE)
}
\arguments{
\item{x, q}{Numeric vectors of quantiles.}

\item{log, log.p}{A logical scalar; if TRUE, probabilities p are given as
log(p).}

\item{p}{A numeric vector of probabilities in [0,1].}

\item{loc, scale, shape}{Numeric vectors.
Location, scale and shape parameters.
All elements of \code{scale} must be positive.}

\item{n}{Numeric scalar.  The number of observations to be simulated.
If \code{length(n) > 1} then \code{length(n)} is taken to be the number
required.}

\item{lower.tail}{A logical scalar.  If TRUE (default), probabilities
are P[X <= x], otherwise, P[X > x].}

\item{m}{A numeric scalar.  The distribution is reparameterised by working
with the GEV(\code{loc, scale, shape}) distribution function raised to the
power \code{m}.  See \strong{Details}.}
}
\description{
Density, distribution function, quantile function and random generator
for the distribution of Fisher's transformation of Pearson's product
moment correlation, based on a random sample from a bivariate normal
distribution
}
\examples{
dFPearson(x = 0, N = 10)
}
\seealso{
\code{\link[SuppDists]{Pearson}} for dpqr functions for the
  untransformed Pearson produce moment correlation coefficient.
}
