% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_menu.R
\name{movies}
\alias{movies}
\title{Main menu for smovie movies}
\usage{
movies(fixed_range = TRUE, hscale = NA, vscale = hscale)
}
\arguments{
\item{fixed_range}{A logical scalar.  Only relevant to the \strong{Discrete}
and \strong{Continuous} menus.  If \code{TRUE} then in the call to
\code{\link{discrete}} or \code{\link{continuous}} the argument
\code{var_support} (\code{discrete}) or \code{var_range}
(\code{continuous}) is set so that the values on the horizontal axes
are fixed at values that enable the movie to show the effects of changing
the parameters of the distribution, at least locally to the default
initial values for the parameters.  For greater control call
\code{\link{discrete}} or \code{\link{continuous}} directly.}

\item{hscale, vscale}{Numeric scalars.  Scaling parameters for the size
of the plot when \code{panel_plot = TRUE}. The default values are 1.4 on
Unix platforms and 2 on Windows platforms.}
}
\description{
Uses the template \code{\link[rpanel]{rp.cartoons}} function to produce
a menu panel from which any of the movies in
\code{\link[smovie]{smovie}} package can be launched. For greater control
of an individual example call the relevant function directly.
}
\examples{
movies()
}
\seealso{
\code{\link{discrete}}, \code{\link{continuous}},
  \code{\link{clt}}, \code{\link{cltq}}, \code{\link{ett}},
  \code{\link{mean_vs_median}}, \code{\link{correlation}},
  \code{\link{lev_inf}}, \code{\link{wws}}, \code{\link{shypo}}.

\code{\link{smovie}}: general information about smovie.
}
