% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_pmf_movie.R
\name{binom_pmf}
\alias{binom_pmf}
\title{Binomial p.m.f. movie}
\usage{
binom_pmf(starting_n = 1, starting_p = 1/2, delta_n = 1, delta_p = 0.05,
  observed_value = NA, ...)
}
\arguments{
\item{starting_n}{A numeric scalar.  The value of n for the first graph.}

\item{starting_p}{A numeric scalar.  The value of p for the first graph.}

\item{delta_n}{A numeric scalar.  The amount by which n is increased
(or decreased) after one click of the + (or -) button in the parameter
window.}

\item{delta_p}{A numeric scalar.  The amount by which p is increased
(or decreased) after one click of the + (or -) button in the parameter
window.}

\item{observed_value}{A non-negative integer.  If \code{observed_value} is
supplied then the corresponding line in the plot of the p.m.f. is coloured
in red.  If \code{observed_value} is not an integer then
\code{round(observed_value)} is used.}

\item{...}{Additional arguments to be passed to
\code{\link[rpanel]{rp.doublebutton}}, not including \code{panel},
\code{variable}, \code{title}, \code{step}, \code{action}, \code{initval},
\code{range}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate how the probability mass function (p.m.f.) of a
binomial (n, p) random variable depends on n and p.
}
\details{
Add details.
}
\examples{
\dontrun{
binom_pmf()

# Increase n and see what happens
binom_pmf(delta_n = 10)

# Highlight observed value of 26 successes in 44 trials
binom_pmf(starting_n = 44, starting_p = 0.1, delta_p = 0.1,
          observed_value = 26)
}
}
\seealso{
\code{\link{smovie}}: general information about smovie.
}
