% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{discrete}
\alias{discrete}
\title{Univariate Discrete Distributions: p.m.f and c.d.f.}
\usage{
discrete(distn, params = list(), plot_par = list(), panel_plot = TRUE,
  hscale = NA, vscale = hscale, observed_value = NA, ...)
}
\arguments{
\item{distn}{Either a character string or a function to choose the discrete
  random variable.

  Strings \code{"binomial"}, \code{"geometric"},
  \code{"hypergeometric"}, \code{"negative binomial"}, \code{Poisson} are
  recognised, case being ignored.  The relevant distributional functions
  \code{dxxx} and \code{pxxx} in the \code{\link[stats]{stats-package}}
  are used.  The abbreviations \code{xxx} are also recognised.
  If \code{distn = "hypergeometric"} then the \code{(size, prob)}
  parameterisation is used, unless a value for \code{mu} is provided
  via the argument \code{params} when the \code{(size, mu)}
  parameterisation is used.

  Valid functions are set up like a standard distributional function
  \code{dxxx}, with first argument \code{x}, last argument \code{log}
  and with arguments to set the parameters of the distribution in between.
  See the \href{https://cran.r-project.org/web/views/Distributions.html}{
  CRAN task view on distributions}.

  If \code{distn} is not supplied then \code{distn = "binomial"}
  is used.}

\item{params}{A named list of initial parameter values with which to start
  the movie.  If \code{distn} is a string and a particular parameter value
  is not supplied then the following values are used.
  \code{"binomial"}: \code{size = 10, prob = 0.5};
  \code{"geometric"}: \code{prob = 0.5};
  \code{"hypergeometric"}: \code{m = 10, n = 7, k = 8};
  \code{"negative binomial"}: \code{size = 10, prob = 0.5};
  \code{"poisson"}: \code{lambda = 5}.

  If \code{distn} is a function then \code{params} must set any required
  parameters.}

\item{plot_par}{A named list of graphical parameters
(see \code{link[graphics]{par}}) to be passed to
\code{\link[graphics]{plot}}.  This may be used to alter the appearance
of the plots of the p.m.f. and c.d.f.}

\item{panel_plot}{A logical parameter that determines whether the plot
is placed inside the panel (\code{TRUE}) or in the standard graphics
window (\code{FALSE}).  If the plot is to be placed inside the panel
then the tkrplot library is required.}

\item{hscale, vscale}{Numeric scalars.  Scaling parameters for the size
of the plot when \code{panel_plot = TRUE}. The default values are 1.4 on
Unix platforms and 2 on Windows platforms.}

\item{observed_value}{A non-negative integer.  If \code{observed_value} is
supplied then the corresponding line in the plot of the p.m.f. is coloured
in red.}

\item{...}{Additional arguments to be passed to
\code{\link[rpanel]{rp.doublebutton}}, not including \code{panel},
\code{variable}, \code{title}, \code{step}, \code{action}, \code{initval},
\code{range}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate how the probability mass function (p.m.f.) and
cumulative distribution function (c.d.f.) of a discrete random variable
depends on the values of its parameters.
}
\details{
The movie starts with a plot of the p.m.f. of the distribution
  for the initial values of the parameters.  There are buttoms to
  increase (+) or decrease (-) each parameter.  There is a radio button
  to switch the plot from the p.m.f. to the c.d.f.

  Owing to a conflict with the argument \code{size} of the function
  \code{\link[rpanel]{rp.control}} the parameter \code{size} of,
  for example, the binomial and negative binomial distributions, is
  labelled as \code{n}.
}
\examples{
\dontrun{
discrete()

discrete(distn = dbinom, params = list(size = 10, prob = 0.5))
}
}
\seealso{
\code{\link{smovie}}: general information about smovie.
}
