% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETT.R
\name{ett}
\alias{ett}
\title{Extremal Types Theorem}
\usage{
ett(n = 30, distn = c("exponential", "uniform", "gp", "normal", "beta"),
  params = list(), panel_plot = TRUE, hscale = NA, vscale = hscale,
  n_add = 1, delta_n = 1, xlab = "x", pos = 1,
  envir = as.environment(pos), ...)
}
\arguments{
\item{n}{An integer scalar.  The size of the samples drawn from the
distribution chosen using \code{distn}.}

\item{distn}{A character scalar specifying the distribution from which
observations are sampled..   Distributions \code{"exponential"},
\code{"uniform"}, \code{"gp"}, \code{"normal"} and \code{"beta"} are
recognised, case being ignored.
The \code{"gp"} case uses the distributional functions
\code{\link[revdbayes]{gp}} in the
\code{\link[revdbayes]{revdbayes-package}}.  The other cases
use the distributional functions in the
\code{\link[stats]{stats-package}}.}

\item{params}{A named list of additional arguments to be passed to the
density function associated with distribution \code{distn}.}

\item{panel_plot}{A logical parameter that determines whether the plot
is placed inside the panel (\code{TRUE}) or in the standard graphics
window (\code{FALSE}).  If the plot is to be placed inside the panel
then the tkrplot library is required.}

\item{hscale, vscale}{Numeric scalars.  Scaling parameters for the size
of the plot when \code{panel_plot = TRUE}. The default values are 1.4 on
Unix platforms and 2 on Windows platforms.}

\item{n_add}{An integer scalar.  The number of simulated datasets to add at
to each new frame of the movie.}

\item{delta_n}{A numeric scalar.  The amount by which n is increased
(or decreased) after one click of the + (or -) button in the parameter
window.}

\item{xlab}{A character scalar.  A name to use to label the horizontal
axis of the plots.}

\item{pos}{A numeric integer.  Used in calls to \code{\link{assign}}
to make information available across successive frames of a movie.
By default, uses the current environment.}

\item{envir}{An alternative way (to \code{pos}) of specifying the
environment. See \code{\link{environment}}.}

\item{...}{Additional arguments to the rpanel functions
\code{\link[rpanel]{rp.button}} and
\code{\link[rpanel]{rp.doublebutton}}, not including \code{panel},
\code{variable}, \code{title}, \code{step}, \code{action}, \code{initval},
\code{range}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate the extremal types theorem, that is, convergence
of the distribution of maxima of random samples from certain distributions
to a member of the Generalized Extreme Value (GEV) family.
}
\details{
Loosely speaking, a consequence of the
  \href{https://en.wikipedia.org/wiki/Extreme_value_theory#Univariate_theory}{Extremal Types Theorem}
  is that, in many situations, the maximum of a \strong{large number} of
  independent random variables has \strong{approximately} a GEV
  distribution.

  Samples of size \code{n} are repeatedly simulated from the distribution
  chosen using \code{distn}.  These samples are summarized using a histogram
  that appears at the top of the movie screen.  For each sample the maximum
  of these \code{n} values is calculated, stored and added to another
  histogram plotted below the first histogram.
  The probability density function (p.d.f.) of the original
  variables is superimposed on the top histogram.  On the bottom histogram
  is superimposed the approximate (large \code{n}) GEV p.d.f. implied by
  the ETT.

  Once it starts, two aspects of this movie are controlled by the user.
  Firstly, there are buttons to increase (+) or decrease (-) the sample
  size, that is, the number of values over which a mean is calculated.
  Then there is a button labelled
  "simulate another \code{n_add} samples of size n".
  Each time this button is clicked \code{n_add} new samples are simulated
  and their sample maxima are added to the bottom histogram.
}
\examples{
\dontrun{
ett()
ett(repeatdelay = 100, repeatinterval = 100)
ett(distn = "uniform")
ett(distn = "gp", params = list(shape = 0.5))
}
}
\seealso{
\code{\link{smovie}}: general information about smovie.
}
