% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_vs_median.R
\name{mean_vs_median}
\alias{mean_vs_median}
\title{Sample mean vs sample median: normally distributed data}
\usage{
mean_vs_median(n = 10, panel_plot = TRUE, hscale = NA, vscale = hscale,
  n_add = 1, delta_n = 1, arrow = TRUE, leg_cex = 1.75, ...)
}
\arguments{
\item{n}{An integer scalar.  The size of the samples drawn from a
standard normal distribution.}

\item{panel_plot}{A logical parameter that determines whether the plot
is placed inside the panel (\code{TRUE}) or in the standard graphics
window (\code{FALSE}).  If the plot is to be placed inside the panel
then the tkrplot library is required.}

\item{hscale, vscale}{Numeric scalars.  Scaling parameters for the size
of the plot when \code{panel_plot = TRUE}. The default values are 1.4 on
Unix platforms and 2 on Windows platforms.}

\item{n_add}{An integer scalar.  The number of simulated datasets to add
to each new frame of the movie.}

\item{delta_n}{A numeric scalar.  The amount by which n is increased
(or decreased) after one click of the + (or -) button in the parameter
window.}

\item{arrow}{A logical scalar.  Should an arrow be included to show the
simulated sample maximum from the top plot being placed into the
bottom plot?}

\item{leg_cex}{The argument \code{cex} to \code{\link[graphics]{legend}}.
Allows the size of the legend to be controlled manually.}

\item{...}{Additional arguments to the rpanel functions
\code{\link[rpanel]{rp.button}} and
\code{\link[rpanel]{rp.doublebutton}}, not including \code{panel},
\code{variable}, \code{title}, \code{step}, \code{action}, \code{initval},
\code{range}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to compare the sampling distributions of the sample mean
and sample median based on a random sample of size \eqn{n} from a
standard normal distribution.
}
\details{
The movie is based on simulating repeatedly samples of size
  \code{n} from a standard normal N(0,1) distribution.  It contains
  three plots.  The top plot contains a histogram of the most recently
  simulated dataset, with the N(0,1) probability density function (p.d.f.)
  superimposed.

  Each time a sample is simulated the sample mean and sample median are
  calculated.  These values are indicated on the top plot as a filled
  circle: a red circle for the sample mean and a blue circle for the
  sample median; and are added to the values plotted in the middle and
  bottom plots.

  The plot in the middle contains a histogram of
  the sample means of \emph{all} the simulated samples.  The p.d.f.s of
  the sampling distributions of the sample mean (thick red line)
  and the sample medians (thin blue line) are superimposed.

  The plot on the bottom contains a histogram of
  the sample medians of \emph{all} the simulated samples.  The p.d.f.s of
  the sampling distributions of the sample mean (thin red line)
  and the sample medians (thick blue line) are superimposed.

  An extra sample is prodcued by clicking the + button next to
  "number of samples of size n". [Ignore the - button.]
  The value of the sample size can be changed using the +/-
  buttons in the panel.
}
\examples{
mean_vs_median()
}
\seealso{
\code{\link{movies}}: a user-friendly menu panel.

\code{\link{smovie}}: general information about smovie.
}
