% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_menu.R
\name{movies}
\alias{movies}
\title{Main menu for smovie movies}
\usage{
movies(hscale = 1)
}
\arguments{
\item{hscale}{Add description.}
}
\description{
Uses the template \code{\link[rpanel]{rp.cartoons}} function to produce
a menu panel from which any of the movies in
\code{\link[smovie]{smovie}} package can be launched.
}
