% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smovie.R
\docType{package}
\name{smovie}
\alias{smovie}
\alias{smovie-package}
\title{smovie: some movies to illustrate concepts in statistics}
\description{
Add description
}
\seealso{
\code{\link{wws}}: Wald, Wilks and Score test movie.

\code{\link{shypo}}: Testing simple hypotheses movie.
}
