% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_wilks_score.R
\name{wws}
\alias{wws}
\title{Wald, Wilks and Score test movie}
\usage{
wws(model = c("norm", "binom"), loglik = NULL, theta_range = c(0.1, 0.7),
  theta0 = mean(theta_range), delta_theta0 = abs(diff(theta_range))/20,
  theta_mle = NULL, alg_score = NULL, alg_obs_info = NULL, digits = 3,
  ...)
}
\arguments{
\item{loglik}{An R function, vectorised with respect to its first argument
that returns the value of the log-likelihood (up to an additive constant).}

\item{theta_range}{A numeric vector of length 2.  The range of values of
\eqn{\theta} over which to plot the log-likelihood.}

\item{theta0}{A numeric scalar.  The value of \eqn{\theta} under the null
hypothesis to use at the start of the movie.}

\item{delta_theta0}{A numeric scalar.  The amount by which the value of
\code{theta0} is increased (or decreased) after one click of the + (or -)
button in the parameter window.}

\item{theta_mle}{A numeric scalar.  The user may use this to supply the
value of the maximum likelihood estimate (MLE) of \eqn{theta}.
Otherwise, \code{\link[stats]{optim}} is used to search for the MLE,
using \code{theta0} as the initial value and \code{theta_range} as
bounds within which to search.}

\item{alg_score}{A R function that returns the score function, that is,
the derivative of \code{loglik} with respect to \eqn{\theta}.}

\item{alg_obs_info}{A R function that returns the observed information
that is, the negated second derivative of \code{loglik} with respect
to \eqn{\theta}.}

\item{...}{Additional arguments to be passed to \code{loglik}.}
}
\value{
Nothing is returned, only the animation is produced.
}
\description{
A movie to illustrate the nature of the Wald, Wilks and score
likelihood-based test statistics, for a model with a scalar unknown
parameter.  The user can change the value of the parameter under a simple
null hypothesis and observe the effect on the test statistics.
}
\details{
Add details.
}
\examples{
# Load package rpanel
# [Use install.packages("rpanel") if necessary]
library(rpanel)

\dontrun{
# Example 6.2 in the STAT3001 notes
wws(theta0 = 0.5)

wws(theta0 = 0.5, model = "binom")

wws(theta0 = 0.5, model = "binom", n_success = 1000, n_failure = 1000,
    theta_range = c(0.4, 0.6))

# data =
# automatic choice of theta_range for binom and norm

set.seed(47)
x <- rnorm(10)
wws(theta0 = 0.2, model = "norm", theta_range = c(-1, 1))

# Log-likelihood for a binomial experiment (up to an additive constant)
bin_loglik <- function(p, n_success, n_failure) {
  return(n_success * log(p) + n_failure * log(1 - p))
}

wws(bin_loglik, theta0 = 0.5, theta_range = c(0.1, 0.7),
    theta_mle = 7 / 20, n_success = 7, n_failure = 13)

bin_alg_score <- function(p, n_success, n_failure) {
  return(n_success / p - n_failure / (1 - p))
}
bin_alg_obs_info <- function(p, n_success, n_failure) {
  return(n_success / p ^ 2 + n_failure / (1 - p) ^ 2)
}
wws(bin_loglik, theta0 = 0.5, theta_range = c(0.1, 0.7),
    theta_mle = 7 / 20, n_success = 7, n_failure = 13,
    alg_score = bin_alg_score, alg_obs_info = bin_alg_obs_info)
}
}
